% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicatorRoutine-package.R
\docType{package}
\name{indicatorRoutine-package}
\alias{indicatorRoutine}
\alias{indicatorRoutine-package}
\title{indicatorRoutine: The Title of the Project}
\description{
A paragraph providing a full description of the project (on several lines...)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MathildeVimont/indicatorRoutine}
  \item Report bugs at \url{https://github.com/MathildeVimont/indicatorRoutine/issues}
}

}
\author{
\strong{Maintainer}: Mathilde Vimont \email{mathilde.vimont@gmail.com} (\href{https://orcid.org/{{orcid}}}{ORCID}) [copyright holder]

}
\keyword{internal}
